requirements.txt
azure-cognitiveservices-speech==1.32.1
python-dotenv==1.0.0
langchain>=0.1.0
numpy>=1.22.0
pyaudio==0.2.13
pytest==7.4.0
typing-extensions>=4.5.0
openai
PyMuPDF